requirements.txt
Flask
Flask_SQLAlchemy
Flask_Bcrypt
python-dotenv
APScheduler
requests
flask-login
psycopg2-binary
postmarker
selenium
webdriver-manager